+++ design-system/tokens/src/artifacts/figma/atlassian-dark/sync-figma-tokens.js
@@ -1173,7 +1173,7 @@ synchronizeFigmaTokens('AtlassianDark', {
       "state": "active",
       "description": "A secondary background for the UI commonly used for grouping items, such as Jira cards in columns."
     },
-    "value": "#03040442"
+    "value": "#101214"
   },
   "Elevation/Surface/Raised": {
     "attributes": {